+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Please mark all change in change log and use the issue from GitHub
 
 ## Bug
 -   \#1705 Limit the insert data batch size
+-   \#1929 Skip MySQL meta schema field width check
 
 ## Feature